+++ src/main/java/com/mongodb/kafka/connect/sink/processor/BlockListValueProjector.java
@@ -18,13 +18,19 @@
 
 package com.mongodb.kafka.connect.sink.processor;
 
+import static com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.FieldProjectionType.BLOCKLIST;
+import static com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.VALUE_PROJECTION_LIST_CONFIG;
+
 import com.mongodb.kafka.connect.sink.MongoSinkTopicConfig;
+import com.mongodb.kafka.connect.sink.processor.field.projection.BlockListProjector;
 
-/** @deprecated use {@link AllowListValueProjector} instead */
-@Deprecated
-public class WhitelistValueProjector extends AllowListValueProjector {
+public class BlockListValueProjector extends BlockListProjector {
 
-  public WhitelistValueProjector(final MongoSinkTopicConfig config) {
-    super(config);
+  public BlockListValueProjector(final MongoSinkTopicConfig config) {
+    this(config, config.getString(VALUE_PROJECTION_LIST_CONFIG));
+  }
+
+  public BlockListValueProjector(final MongoSinkTopicConfig config, final String fieldList) {
+    super(config, buildProjectionList(BLOCKLIST, fieldList), SinkDocumentField.VALUE);
   }
 }