+++ src/main/java/com/mongodb/kafka/connect/sink/processor/id/strategy/PartialValueStrategy.java
@@ -18,20 +18,30 @@
 
 package com.mongodb.kafka.connect.sink.processor.id.strategy;
 
+import static com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.DOCUMENT_ID_STRATEGY_CONFIG;
+import static com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.FieldProjectionType.ALLOWLIST;
+import static com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.FieldProjectionType.BLOCKLIST;
+import static com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.VALUE_PROJECTION_LIST_CONFIG;
+import static com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.VALUE_PROJECTION_TYPE_CONFIG;
+import static java.lang.String.format;
+
 import org.apache.kafka.connect.sink.SinkRecord;
 
 import org.bson.BsonDocument;
 import org.bson.BsonValue;
 
+import com.mongodb.kafka.connect.sink.MongoSinkTopicConfig;
+import com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.FieldProjectionType;
 import com.mongodb.kafka.connect.sink.converter.SinkDocument;
+import com.mongodb.kafka.connect.sink.processor.AllowListValueProjector;
+import com.mongodb.kafka.connect.sink.processor.BlockListValueProjector;
 import com.mongodb.kafka.connect.sink.processor.field.projection.FieldProjector;
+import com.mongodb.kafka.connect.util.ConnectConfigException;
 
 public class PartialValueStrategy implements IdStrategy {
   private FieldProjector fieldProjector;
 
-  public PartialValueStrategy(final FieldProjector fieldProjector) {
-    this.fieldProjector = fieldProjector;
-  }
+  public PartialValueStrategy() {}
 
   @Override
   public BsonValue generateId(final SinkDocument doc, final SinkRecord orig) {
@@ -40,12 +50,38 @@ public class PartialValueStrategy implements IdStrategy {
     // happening later in the chain e.g. for value fields
     SinkDocument clone = doc.clone();
     fieldProjector.process(clone, orig);
-    // NOTE: If there is no key doc present the strategy
+    // NOTE: If there is no value doc present the strategy
     // simply returns an empty BSON document per default.
     return clone.getValueDoc().orElseGet(BsonDocument::new);
   }
 
   public FieldProjector getFieldProjector() {
     return fieldProjector;
+  }
+
+  @Override
+  public void configure(final MongoSinkTopicConfig config) {
+    FieldProjectionType valueProjectionType =
+        FieldProjectionType.valueOf(config.getString(VALUE_PROJECTION_TYPE_CONFIG).toUpperCase());
+
+    String fieldList = config.getString(VALUE_PROJECTION_LIST_CONFIG);
+
+    switch (valueProjectionType) {
+      case BLACKLIST:
+      case BLOCKLIST:
+        fieldProjector = new BlockListValueProjector(config, fieldList);
+        break;
+      case ALLOWLIST:
+      case WHITELIST:
+        fieldProjector = new AllowListValueProjector(config, fieldList);
+        break;
+      default:
+        throw new ConnectConfigException(
+            DOCUMENT_ID_STRATEGY_CONFIG,
+            this.getClass().getName(),
+            format(
+                "Invalid %s value. It should be set to either %s or %s",
+                VALUE_PROJECTION_TYPE_CONFIG, BLOCKLIST, ALLOWLIST));
+    }
   }
 }